requirements.txt
jsonpickle==1.2
pkg-resources==0.0.0
pycrypto==2.6.1
